Remove conflicting image files
If an image file already exists on the system when trying to write a new
file, it's an indication that a previous image write went wrong
somewhere. So instead of failing when the image already exists, we
should remove the image.

This is technically still racy, but in that scenario the error behavior
is fine.